+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Support for custom fields. These fields allow you to run custom filter
   functions for anything that cannot be expressed with Flop filters.
-- Added `Flop.with_named_bindings/3` for dynamically adding bindings needed for
+- Added `Flop.with_named_bindings/4` for dynamically adding bindings needed for
   a Flop query.
 - Added `fetch`, `get`, `get_all`, `delete`, `delete_first`, `drop`, `new`,
   `take`, `pop`, `pop_first`, `put` and `put_new` functions to `Flop.Filter`.